Fixed heating and sharpening GetOffsetPercentages.

+++ GGJ20/Assets/Scripts/Work/Data/Tasks/HeatingTaskScriptableObject.cs
@@ -4,7 +4,7 @@
 public class HeatingTaskScriptableObject : TaskScriptableObject
 {
     [Range(0.0f, 1.0f)] 
-    public int targetHeatPercentage;
+    public float targetHeatPercentage;
     public bool HeatMatters;
 
     public override WorkManager.TaskType GetTaskType()
